requirements.txt
python-dotenv 
pandas
numpy
scikit-learn
joblib
pymongo
certifi
pymongo[srv]

-e .